Add git hook to check commit message formatting
A proper commit message should have a subject, followed by a blank
line, and then at least one line of body. If any commit message does
not conform to that, the commit will be rejected and the user needs to
try again.

It allows for lines (except subject) to be longer than 72
characters, but warns about it.