Add Scripts asset pipeline
== Bower ==
Bower can be used to pull in full libraries, where what is used depends
on the particular library.

== Modernizr ==
Modernizr should be as little as possible, so before concatenating, the
grunt-modernizr task runs, creating a non-uglified dev version with all
tests used in the JavaScript or .scss files.

== Concatenation ==
The build targets are specified in grunt/tasks/contrib-concat.js, and
that's where other files/targets should be added. It does not support
EmberJS -- despite that being used in the current master --, as I need
to reconsider the pipeline for EmberJS.

== Uglify ==
It does not mangle any of the symbols, but removes superfluous
whitespace and similar things, outputting as small a files as possible.