Add deploy task for HTML
The deploy task for HTML builds the various HTML-assets, before it
syncs the folders (../docs for local, /var/www/docs.exaltedage.net for
production), deleting any files that are not in the build. NB: Ensure
that any asset files are also kept on the local. Should not be an issue
for the docs, however.

Commands:
- grunt deploy:html (assumes local)
- grunt deploy:html:production (production)
- grunt deploy:html:local (local)

Fix #9.