Fix #15: Markdown not working
It was not due to Markdown. The issue is that Purifier doesn't seem to
read the proper configuration. The documentation claims that it should
be reading the file `app/config/packages/mews/purifier/config.php`, but
it doesn't.

Instead, I kept the location to `app/config/purifier` and loaded that
in the Markdown macro.